Split runtime error diagnostics
- Keep ExecutionError root constructors in runtime/error.rs.
- Move source panic data and behavior into runtime/error/panic.rs.
- Move miette diagnostic rendering and failed-value formatting into runtime/error/diagnostic.rs.
- Preserve existing runtime behavior and execution-error fixture output.
